Designate data source and delegate for listings.
The delegate for the listings UITableView is its controlling
ASListingsViewController. The data source is an object of new type
ASQueryController.

Since UITableView.delegate is an assign property, the ASQueryController
is not retained. This causes crashes if the ASQueryController is
destroyed. The strong reference to the delegate from
ASListingsViewController prevents it from being unexpectedly destroyed.