Create data source class for a list of businesses.
ASBusinessList represents a list of businesses and is to be used as the
data source for the listings table view.  The contained objects should
be accessed using KVC accessors.

As the eventual data source for the listings table view, ASBusinessList
implements the mandatory methods of the UITableViewDataSource protocol.
The current implementation of tableView:cellForRowAtIndexPath: returns
a very basic cell.