Create controller for business list class.
ASBusinessListDataController acts as the proxy between the local model
objects and the API. It maintains a strong reference to a single
instance of ASBusinessList. When it receives an updateData: message, it
queries the API and begins downloading the response asynchronously. When
the response is fully downloaded, it is parsed and used to initialize
a new copy of ASBusinessList, which replaces the old one.