Have listings table VC observe its business list.
The listings table view controller owns its own instance of the business
list data controller, which in turn controls a list of businesses. This
list is also the data source for the listings table view.

The view controller observes the "businesses" key of the data
controller. When the data controller changes the value of that property
(usually from downloading fresh data), the view controller is notified.
It assigns the new list as the table view's data source and refreshes
the table view.