Make the listings table fetch new data on launch.
Eventually, the listings table view controller will only fetch new data
in response to a specific set of stimuli, such as the user actively
submitting a query. However, when its view is initially displayed, it
probably should show something. When the table view is set to appear but
has no data source, the view controller initiates an update
automatically.