Fix broken API access.
The URL has changed from "/ios" to "/api", and several of the JSON
fields have been renamed.

Previously, the app crashed if fetching the JSON failed, because the KVO
change dictionary would return NSNull instead of nil. We now avoid this
by explicitly setting the listing table's data source to nil in that
situation.