Make business properties single-assignment.
It's not really feasible to make ASBusiness immutable, since that would
require a huge initializer. Next best thing is probably to make each
property single-assignment-only, throwing an exception if additional
assignments are attempted.

Each property has a custom setter now. The setter checks whether the
backing ivar is nil; if it is, it allows the assignment. Otherwise, it
throws an exception.

Conflicts:
	AllSortz.xcodeproj/project.pbxproj