Add business ID property to ASListing.
Since the push segue that brings up the details view will need to pass
the ID of the selected business to the details VC, the list of
businesses needs to keep track of IDs.

We require that the ID be provided on init. Having a business proxy
object without an ID is nonsensical.