Reviews can be added to the business.
There is also a way to now modify the actual content associated with a business.
This might night be functionality we want to expose to the users for the demo,
but it's something that we can use internally. Also, the primary view is now the
map view. Thus, the listing does nothing but just dump what you see on the map
view. Thus, there's no "load more rows" as you scroll. The query / search
functionality remains the same for now.

There is still some trickiness with getting the 'zoom to fit annotations'
behavior to work as expected. On the one hand, you don't want the application
randomly resizing. On the other hand, you might want to resize so you can see
all the pins