MAJOR code refactoring
Remove ASListing since it was essentially the same as ASBusiness
(different fields are used in different views, but the same overall
data)

IconDownloader has been factored into the businesslist since that's the
only place that it should be used.

I removed most of the Query files since the query is only textual right
now. There's no sense in keeping the controllers around for a fancy
view.

I also tried to reorganize the code as much as possible so that it is
logically spread out

AddBusiness and EditBusiness need a lot of work, but I think that a lot
what goes into these views needs design thought anyway

Added the property TestFlight token