Refactoring and fix add business topic
Still trying to refactor the code as much as possible. Removed
unecessary outlets, and changed the name of the URLencoding file to
global since it contains things used throughout the app.

The navigation bar color is now set in the app delegate at app startup

The UserProfile page is working much more smoothly. I had done some
weird stuff before that was making it not dependable at all.

Made a singleton for the device controller. This is important. Now,
there's a single object that controls the location as well as user info.
We should start migrating stuff into core data at some point as well.
The device "manager" still defines a delegate for the location update.
This is because the device manager itself cannot know when the location
is going to be updated. Views (like map view) need to do certain things
when the location is updated, which makes the delegate necessary.

There's some locks that I use in the map controller to ensure that not
too many updates are sent to the server. We might want to look into this
to see if there's a  way to smooth it out.

Finally, I updated the add business topic to be more stable and not
instant-select something when a user scrolls to it